Add postgresql role
In order for PostgreSQL to properly create a database the operating system's locale has to be UTF-8. An additional role, os, was created to fulfill this purpose while opening the door for other common "os" level changes. For consistency this is also applied to plays that use an "os".

Also bumping up base LXC image to Debian 13 (Trixie). This turned out to be a huge mistake as fresh installations of 'ca' and 'dns' role now break. Various changes to both roles and dependencies was needed to make this all function correctly.